refactor: 💡 re-organised generation code
