+++ internal/generator/gen_posts.go
@@ -1,0 +1,32 @@
+package generator
+
+import (
+	"html/template"
+	"path/filepath"
+	"time"
+
+	"github.com/intothevoid/likho/internal/config"
+	"github.com/intothevoid/likho/internal/parser"
+	"github.com/intothevoid/likho/internal/post"
+)
+
+func generateAllPostsHTML(cfg *config.Config, tmpl *template.Template, posts []post.Post, pages []parser.Page) error {
+	data := struct {
+		Posts       []post.Post
+		SiteTitle   string
+		CurrentYear int
+		PageTitle   string
+		Content     template.HTML
+		Pages       []parser.Page
+	}{
+		Posts:       posts,
+		SiteTitle:   cfg.Site.Title,
+		CurrentYear: time.Now().Year(),
+		PageTitle:   "Posts",
+		Content:     "", // Leave empty as we're not using it directly
+		Pages:       pages,
+	}
+
+	outputPath := filepath.Join(cfg.Content.OutputDir, "posts.html")
+	return executeTemplate(tmpl, "posts.html", outputPath, data)
+}
